requirements.txt
polars
scipy
joblib
scikit-learn
requests
cryptography
python-dateutil
websockets
aiohttp
sortedcontainers
pymc
arviz
